Fix compilation error with GCC 9 + LTO on 64bit
Fix "lto-type-mismatch" error raised when size_t is defined as
unsigned long and not unsigned int. This is the case e.g. with
GCC 9 on x86_64.

+++ src/match_openssh.c
@@ -150,7 +150,7 @@ static EVP_PKEY *ssh1_line_to_key(char *line)
 	return NULL;
 }
 
-extern int sc_base64_decode(const char *in, unsigned char *out, size_t outlen);
+extern int sc_base64_decode(const char *in, unsigned char *out, unsigned int outlen);
 
 static EVP_PKEY *ssh2_line_to_key(char *line)
 {
